+++ src/tekton/triggertype.ts
@@ -3,8 +3,7 @@
  *  Licensed under the MIT License. See LICENSE file in the project root for license information.
  *-----------------------------------------------------------------------------------------------*/
 
-import { ItemPath, Params, VCT } from './pipelinecontent';
-import { VolumeTypeSecret, VolumeTypeConfigMaps, VolumeTypePVC, PipelineRunWorkspace } from '../tekton';
+import { VolumeTypeSecret, VolumeTypeConfigMaps, VolumeTypePVC, PipelineRunWorkspace, Params, VCT, ItemPath } from '../tekton';
 
 export interface TriggerFormValues {
   selectedTrigger: string;
