+++ test/tekton/pipeline.test.ts
@@ -14,9 +14,9 @@ import { PipelineExplorer } from '../../src/pipeline/pipelineExplorer';
 import { TektonItem } from '../../src/tekton/tektonitem';
 import { TestItem } from './testTektonitem';
 import * as vscode from 'vscode';
-import { Trigger, StartObject, NameType, Resources, Params, PipelineContent } from '../../src/tekton/pipelinecontent';
 import { ContextType } from '../../src/context-type';
 import { Command } from '../../src/cli-command';
+import { NameType, Params, Resources, StartObject, Trigger } from '../../src/tekton';
 
 const expect = chai.expect;
 chai.use(sinonChai);
@@ -153,7 +153,7 @@ suite('Tekton/Pipeline', () => {
     });
 
     test('starts a pipeline with appropriate resources', async () => {
-      sandbox.stub(PipelineContent, 'startObject').withArgs(pipeTrigger, 'Pipeline').resolves(startPipelineObj);
+      // sandbox.stub(PipelineContent, 'startObject').withArgs(pipeTrigger, 'Pipeline').resolves(startPipelineObj);
       sandbox.stub(Pipeline, 'start').withArgs(pipelineItem).resolves('Pipeline \'pipeline\' successfully created');
       const result = await Pipeline.start(pipelineItem);
       expect(result).equals(`Pipeline '${startPipelineObj.name}' successfully created`);
